# back/pkg/docs/swagger.yaml
basePath: /api
...
info:
  contact: {}
  description: Documentación de la API de AUTH.
  version: "1.0"
paths:
...
  /get-users/{groupID}:
    get:
      consumes:
      - application/json
      description: Get a list of users belonging to the specified group using the
        provided groupID.
      parameters:
      - description: Group ID
        in: path
        name: groupID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request - invalid groupID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Retrieve users for a group
      tags:
      - Users
      - Groups
...
swagger: "2.0"